lars programm

+++ Lars/Lars programm/Lars programm/Program.cs
@@ -7,7 +7,7 @@ namespace Calculator
         static void Main(string[] args)
         {
             // Declare variables and then initialize to zero.
-            int num1 = 0; int num2 = 0;
+            double num1 = 0; double num2 = 0;
 
             // Display title as the C# console calculator app.
             Console.WriteLine("Console Calculator in C#\r");
@@ -15,11 +15,11 @@ namespace Calculator
 
             // Ask the user to type the first number.
             Console.WriteLine("Zahl eingeben, dann Enter drücken...");
-            num1 = Convert.ToInt32(Console.ReadLine());
+            num1 = Convert.ToDouble(Console.ReadLine());
 
             // Ask the user to type the second number.
             Console.WriteLine("Noch eine Zahl eingeben, dann Enter drücken...");
-            num2 = Convert.ToInt32(Console.ReadLine());
+            num2 = Convert.ToDouble(Console.ReadLine());
 
             // Ask the user to choose an option.
             Console.WriteLine("Wähle eine der folgenden Optionen:");
